+++ plugins/admin-system/src/Admin/Effects/Drunk.cpp
@@ -24,7 +24,7 @@ static constexpr float ShakeAmplitude = 8.0f;
 Effect MakeDrunk(VoltMod::Runtime& runtime)
 {
     return Effect{.Permission = Permission::Fun,
-                  .Id = static_cast<int>(EffectId::Drunk),
+                  .Id = EffectId::Drunk,
                   .NameKey = "action.drunk",
                   .OnKey = "broadcast.drunkOn",
                   .OffKey = "broadcast.drunkOff",
