+++ plugins/admin-system/src/Admin/Effects/Wallhack.cpp
@@ -12,7 +12,7 @@ using VoltMod::GlowVision;
 Effect MakeWallhack(VoltMod::Runtime& runtime)
 {
     return Effect{.Permission = Permission::Wallhack,
-                  .Id = static_cast<int>(EffectId::Wallhack),
+                  .Id = EffectId::Wallhack,
                   .NameKey = "action.wallhack",
                   .OnKey = "broadcast.wallhackOn",
                   .OffKey = "broadcast.wallhackOff",
